Humanize option names
