+++ README.md
@@ -57,15 +57,15 @@ conforms to a consistent style. (See this [blog post](http://jlongster.com/A-Pre
     + [Custom Parser API](#custom-parser-api)
   * [Excluding code from formatting](#excluding-code-from-formatting)
 - [Options](#options)
-  * [bracket-spacing](#bracket-spacing)
-  * [jsx-bracket-same-line](#jsx-bracket-same-line)
-  * [parser](#parser)
-  * [print-width](#print-width)
-  * [semi](#semi)
-  * [single-quote](#single-quote)
-  * [tab-width](#tab-width)
-  * [trailing-comma](#trailing-comma)
-  * [use-tabs](#use-tabs)
+  * [Bracket spacing](#bracket-spacing)
+  * [Jsx bracket same line](#jsx-bracket-same-line)
+  * [Parser](#parser)
+  * [Print width](#print-width)
+  * [Semi](#semi)
+  * [Single quote](#single-quote)
+  * [Tab width](#tab-width)
+  * [Trailing comma](#trailing-comma)
+  * [Use tabs](#use-tabs)
 - [Configuration File](#configuration-file)
   * [Basic Configuration](#basic-configuration)
   * [Configuration Overrides](#configuration-overrides)
@@ -562,7 +562,7 @@ matrix(
 Prettier ships with a handful of customizable format options, usable in both the CLI and API.
 
 <!-- AUTO-GENERATED-CONTENT:START (PRETTIER_OPTIONS:headingLevel=3) -->
-### bracket-spacing
+### Bracket spacing
 
 Print spaces between brackets.
 
@@ -571,7 +571,7 @@ Default | CLI Override | API Override
 --------|--------------|-------------
 `true` | `--bracket-spacing` | `bracketSpacing: <bool>`
 
-### jsx-bracket-same-line
+### Jsx bracket same line
 
 Put > on the last line instead of at a new line.
 
@@ -580,7 +580,7 @@ Default | CLI Override | API Override
 --------|--------------|-------------
 `false` | `--jsx-bracket-same-line` | `jsxBracketSameLine: <bool>`
 
-### parser
+### Parser
 
 Which parser to use.
 
@@ -605,7 +605,7 @@ Default | CLI Override | API Override
 --------|--------------|-------------
 `"babylon"` | <code>--parser &lt;flow&#124;babylon&#124;typescript&#124;css&#124;less&#124;scss&#124;json&#124;graphql&gt;</code> | <code>parser: &lt;flow&#124;babylon&#124;typescript&#124;css&#124;less&#124;scss&#124;json&#124;graphql&gt;</code>
 
-### print-width
+### Print width
 
 The line length where Prettier will try wrap.
 
@@ -627,7 +627,7 @@ Default | CLI Override | API Override
 --------|--------------|-------------
 `80` | `--print-width <int>` | `printWidth: <int>`
 
-### semi
+### Semi
 
 Print semicolons.
 
@@ -636,7 +636,7 @@ Default | CLI Override | API Override
 --------|--------------|-------------
 `true` | `--semi` | `semi: <bool>`
 
-### single-quote
+### Single quote
 
 Use single quotes instead of double quotes.
 
@@ -650,7 +650,7 @@ Default | CLI Override | API Override
 --------|--------------|-------------
 `false` | `--single-quote` | `singleQuote: <bool>`
 
-### tab-width
+### Tab width
 
 Number of spaces per indentation level.
 
@@ -659,7 +659,7 @@ Default | CLI Override | API Override
 --------|--------------|-------------
 `2` | `--tab-width <int>` | `tabWidth: <int>`
 
-### trailing-comma
+### Trailing comma
 
 Print trailing commas wherever possible when multi-line.
 
@@ -674,7 +674,7 @@ Default | CLI Override | API Override
 --------|--------------|-------------
 `"none"` | <code>--trailing-comma &lt;none&#124;es5&#124;all&gt;</code> | <code>trailingComma: &lt;none&#124;es5&#124;all&gt;</code>
 
-### use-tabs
+### Use tabs
 
 Indent with tabs instead of spaces.
