+++ docs/options.md
@@ -6,7 +6,7 @@ title: Options
 Prettier ships with a handful of customizable format options, usable in both the CLI and API.
 
 <!-- AUTO-GENERATED-CONTENT:START (PRETTIER_OPTIONS:headingLevel=2) -->
-## bracket-spacing
+## Bracket spacing
 
 Print spaces between brackets.
 
@@ -15,7 +15,7 @@ Default | CLI Override | API Override
 --------|--------------|-------------
 `true` | `--bracket-spacing` | `bracketSpacing: <bool>`
 
-## jsx-bracket-same-line
+## Jsx bracket same line
 
 Put > on the last line instead of at a new line.
 
@@ -24,7 +24,7 @@ Default | CLI Override | API Override
 --------|--------------|-------------
 `false` | `--jsx-bracket-same-line` | `jsxBracketSameLine: <bool>`
 
-## parser
+## Parser
 
 Which parser to use.
 
@@ -49,7 +49,7 @@ Default | CLI Override | API Override
 --------|--------------|-------------
 `"babylon"` | <code>--parser &lt;flow&#124;babylon&#124;typescript&#124;css&#124;less&#124;scss&#124;json&#124;graphql&gt;</code> | <code>parser: &lt;flow&#124;babylon&#124;typescript&#124;css&#124;less&#124;scss&#124;json&#124;graphql&gt;</code>
 
-## print-width
+## Print width
 
 The line length where Prettier will try wrap.
 
@@ -71,7 +71,7 @@ Default | CLI Override | API Override
 --------|--------------|-------------
 `80` | `--print-width <int>` | `printWidth: <int>`
 
-## semi
+## Semi
 
 Print semicolons.
 
@@ -80,7 +80,7 @@ Default | CLI Override | API Override
 --------|--------------|-------------
 `true` | `--semi` | `semi: <bool>`
 
-## single-quote
+## Single quote
 
 Use single quotes instead of double quotes.
 
@@ -94,7 +94,7 @@ Default | CLI Override | API Override
 --------|--------------|-------------
 `false` | `--single-quote` | `singleQuote: <bool>`
 
-## tab-width
+## Tab width
 
 Number of spaces per indentation level.
 
@@ -103,7 +103,7 @@ Default | CLI Override | API Override
 --------|--------------|-------------
 `2` | `--tab-width <int>` | `tabWidth: <int>`
 
-## trailing-comma
+## Trailing comma
 
 Print trailing commas wherever possible when multi-line.
 
@@ -118,7 +118,7 @@ Default | CLI Override | API Override
 --------|--------------|-------------
 `"none"` | <code>--trailing-comma &lt;none&#124;es5&#124;all&gt;</code> | <code>trailingComma: &lt;none&#124;es5&#124;all&gt;</code>
 
-## use-tabs
+## Use tabs
 
 Indent with tabs instead of spaces.
